Added the sorting algorithm
As expected, this will break some of the unit tests for the helper
methods of this class that require the heapSize variable to be set.
Since the Sort method decrements this variable for each iteration of the
loop, we need to find another way to set this variable only when we are
in the unit test.